+++ docs/api/trace.span.rst
@@ -1,0 +1,7 @@
+opentelemetry.trace.span
+========================
+
+.. automodule:: opentelemetry.trace.span
+    :members:
+    :undoc-members:
+    :show-inheritance:
