+++ ext/opentelemetry-ext-aiohttp-client/src/opentelemetry/ext/aiohttp_client/__init__.py
@@ -145,7 +145,7 @@ def create_trace_config(
         )
 
         trace_config_ctx.token = context_api.attach(
-            trace.propagation.set_span_in_context(trace_config_ctx.span)
+            trace.set_span_in_context(trace_config_ctx.span)
         )
 
         propagators.inject(type(params.headers).__setitem__, params.headers)
