+++ opentelemetry-api/src/opentelemetry/trace/propagation/__init__.py
@@ -13,22 +13,40 @@
 # limitations under the License.
 from typing import Optional
 
-from opentelemetry import trace as trace_api
 from opentelemetry.context import get_value, set_value
 from opentelemetry.context.context import Context
+from opentelemetry.trace.span import INVALID_SPAN, Span
 
 SPAN_KEY = "current-span"
 
 
 def set_span_in_context(
-    span: trace_api.Span, context: Optional[Context] = None
+    span: Span, context: Optional[Context] = None
 ) -> Context:
+    """Set the span in the given context.
+
+    Args:
+        span: The Span to set.
+        context: a Context object. if one is not passed, the
+            default current context is used instead.
+    """
     ctx = set_value(SPAN_KEY, span, context=context)
     return ctx
 
 
-def get_span_from_context(context: Optional[Context] = None) -> trace_api.Span:
+def get_current_span(context: Optional[Context] = None) -> Optional[Span]:
+    """Retrieve the current span.
+
+    Args:
+        context: A Context object. If one is not passed, the
+            default current context is used instead.
+
+    Returns:
+        The Span set in the context if it exists. None otherwise.
+    """
     span = get_value(SPAN_KEY, context=context)
-    if not isinstance(span, trace_api.Span):
-        return trace_api.INVALID_SPAN
+    if span is None:
+        return None
+    if not isinstance(span, Span):
+        return INVALID_SPAN
     return span
