+++ opentelemetry-api/tests/propagators/test_global_httptextformat.py
@@ -17,10 +17,7 @@ import unittest
 
 from opentelemetry import correlationcontext, trace
 from opentelemetry.propagators import extract, inject
-from opentelemetry.trace.propagation import (
-    get_span_from_context,
-    set_span_in_context,
-)
+from opentelemetry.trace import get_current_span, set_span_in_context
 
 
 def get_as_list(
@@ -52,7 +49,7 @@ class TestDefaultGlobalPropagator(unittest.TestCase):
         correlations = correlationcontext.get_correlations(context=ctx)
         expected = {"key1": "val1", "key2": "val2"}
         self.assertEqual(correlations, expected)
-        span_context = get_span_from_context(context=ctx).get_context()
+        span_context = get_current_span(context=ctx).get_context()
 
         self.assertEqual(span_context.trace_id, self.TRACE_ID)
         self.assertEqual(span_context.span_id, self.SPAN_ID)
